v16.3.3 build 2 NOTES - Added: 5 new Player Quests - Added: Back up feature (backs up all data into "plugins\RandomPackage_backups" folder, default is every 360 minutes [6 hours]) - Updated: CollectionFilter - Updated: "collection filter.yml", "config.yml" - Fixed: CustomCreepers not working - Fixed: Not being able to claim Global Challenge Prizes - Fixed: AuctionHouse incorrect auction amount if larger than max double (2.14 million) - Fixed: Titles being unobtainable using the "title:<#>" format - Fixed: Errors upon plugin disable relating to Kits - Fixed: 1 Lootbox error

+++ randompackage-spigot/src/main/java/me/randomhashtags/randompackage/addons/objects/AuctionedItem.java
@@ -2,15 +2,16 @@ package me.randomhashtags.randompackage.addons.objects;
 
 import org.bukkit.inventory.ItemStack;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public class AuctionedItem {
     public long auctionTime;
     public final UUID auctioner;
     private final ItemStack item;
-    public final double price;
+    public final BigDecimal price;
     public boolean claimable;
-    public AuctionedItem(long auctionTime, UUID auctioner, ItemStack item, double price) {
+    public AuctionedItem(long auctionTime, UUID auctioner, ItemStack item, BigDecimal price) {
         this.auctionTime = auctionTime;
         this.auctioner = auctioner;
         this.item = item;
